# contracts/http/openapi.yaml
openapi: 3.0.3
info:
  title: Go Backend Architecture Template API
  version: 0.1.0
  description: |
    HTTP contract for the backend template. Documents the purpose of each API and
    the meaning of request/response fields for frontends and API consumers.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: getHealth
      summary: Check service liveness or readiness
      description: |
        Purpose: used by load balancers and orchestrators to decide if the
        instance can receive traffic. Default mode is ready (all dependencies
        checked); use check=live to only verify the process is running.
      parameters:
        - name: check
          in: query
          required: false
          description: Which check to run. live = process only; ready = process + DB, cache, KV, vector store.
          schema:
            type: string
            enum:
              - live
              - ready
            default: ready
      responses:
        "200":
          description: The requested health check passed; dependencies match the requested mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "400":
          description: Query parameters failed validation (e.g. invalid check value).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: Ready check failed; at least one dependency is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "504":
          $ref: "#/components/responses/RequestTimeout"

  /health/stream:
    get:
      operationId: streamHealth
      summary: Stream bounded health updates (SSE demo)
      description: |
        Demonstrates Server-Sent Events using the existing health check. Sends an
        initial `health` event immediately, then repeats it at the configured
        interval, with keep-alive comments between updates. The stream ends after
        its configured maximum duration or when the client disconnects. It is for
        observing health changes, not for load-balancer or orchestrator probes.
      parameters:
        - name: check
          in: query
          required: false
          description: Which check to run for each event. live = process only; ready = process + DB, cache, KV, vector store.
          schema:
            type: string
            enum:
              - live
              - ready
            default: ready
      responses:
        "200":
          description: A text/event-stream response containing `health` events whose JSON data matches HealthResponse.
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Event frames. Each `health` event's JSON data is a HealthResponse; comments are keep-alives.
              example: |
                event: health
                data: {"database":{"status":"up","name":"app","in_recovery":false,"uptime_seconds":99},"cache":{"status":"up"},"kvstore":{"status":"up"},"vectorstore":{"status":"up"}}

                : keep-alive
        "400":
          description: Query parameters failed validation before the stream opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"

  /auth/otp/send:
    post:
      operationId: sendOTP
      summary: Send a one-time password to an email address
      description: |
        Generates a numeric OTP code and sends it to the provided email.
        Does not reveal whether the email is already registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPSendRequest"
      responses:
        "200":
          description: OTP code sent (or silently ignored if rate-limited).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request body (e.g. missing or malformed email).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/RequestTimeout"

  /auth/otp/verify:
    post:
      operationId: verifyOTP
      summary: Verify an OTP code and create a session
      description: |
        Validates the OTP code for the given email. On success, creates an
        authenticated session and sets a session cookie. If no account exists
        for this email, one is created automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPVerifyRequest"
      responses:
        "200":
          description: OTP verified; session cookie set in response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "401":
          description: OTP code invalid or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/RequestTimeout"

  /auth/oauth/{provider}/authorize:
    get:
      operationId: oauthAuthorize
      summary: Redirect to OAuth provider for authorization
      description: |
        Generates a one-time OAuth state token and a short-lived, HttpOnly
        browser binding cookie (both valid for 10 minutes), then redirects the
        browser to the provider's authorization page with PKCE. The callback
        succeeds only in the browser that initiated this request. Open this URL
        via navigation — a link, `window.location`, or popup — not via
        `fetch`/XHR. The provider parameter selects which OAuth integration to
        use (e.g. google).
      parameters:
        - name: provider
          in: path
          required: true
          description: OAuth provider name (e.g. google).
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OAuth provider's authorization page.
        "400":
          description: Unsupported provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/RequestTimeout"

  /auth/oauth/{provider}/callback:
    get:
      operationId: oauthCallback
      summary: Handle OAuth provider callback
      description: |
        Receives the authorization code and state from the OAuth provider.
        Exchanges the code for user info, finds or creates the account,
        creates a session, and sets a session cookie. The callback also
        requires the short-lived browser binding cookie set by the authorize
        endpoint; it is cleared after this request.
      parameters:
        - name: provider
          in: path
          required: true
          description: OAuth provider name matching the authorize request.
          schema:
            type: string
        - name: code
          in: query
          required: true
          description: Authorization code returned by the OAuth provider.
          schema:
            type: string
        - name: state
          in: query
          required: true
          description: CSRF state token for validation.
          schema:
            type: string
      responses:
        "200":
          description: OAuth login successful; session cookie set in response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing or invalid parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "401":
          description: OAuth state or browser binding is invalid or expired, or code exchange with the provider failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/RequestTimeout"

  /auth/logout:
    post:
      operationId: logout
      summary: End the current session
      description: |
        Revokes the server-side session and clears the session cookie.
        If no session exists, succeeds silently.
      responses:
        "200":
          description: Session ended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/RequestTimeout"

  /auth/me:
    get:
      operationId: getMe
      summary: Get current session info
      description: |
        Returns the authenticated user's account ID and email.
        Requires a valid session cookie.
      responses:
        "200":
          description: Current session info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated or session expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          $ref: "#/components/responses/InternalError"
        "504":
          $ref: "#/components/responses/RequestTimeout"

components:
  responses:
    InternalError:
      description: An unexpected server failure prevented the operation from completing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    RequestTimeout:
      description: The server-side request deadline expired before the operation completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"

  schemas:
    HealthResponse:
      description: Aggregate status of the service and its dependencies for health checks.
      type: object
      required:
        - database
        - cache
        - kvstore
        - vectorstore
      properties:
        database:
          $ref: "#/components/schemas/HealthDatabase"
        cache:
          $ref: "#/components/schemas/HealthDependency"
        kvstore:
          $ref: "#/components/schemas/HealthDependency"
        vectorstore:
          $ref: "#/components/schemas/HealthDependency"
    HealthDatabase:
      description: Primary database health; used to decide if the instance can serve read/write traffic.
      type: object
      required:
        - status
        - name
        - in_recovery
        - uptime_seconds
      properties:
        status:
          description: ok when connected and usable; otherwise indicates failure.
          type: string
        name:
          description: Database identifier (e.g. connection name or role).
          type: string
        in_recovery:
          description: true if the DB is in recovery (e.g. standby); may affect readiness.
          type: boolean
        uptime_seconds:
          description: How long the database has been up; useful for monitoring.
          type: integer
          format: int64
    HealthDependency:
      description: Generic dependency status for cache, KV, and vector store.
      type: object
      required:
        - status
      properties:
        status:
          description: ok when the dependency is reachable and usable; otherwise indicates failure.
          type: string
    APIError:
      description: Standard error payload returned when a request is invalid or the server cannot fulfill it.
      type: object
      required:
        - code
        - message
      properties:
        code:
          description: Machine-readable error code for client handling (e.g. validation, not_found).
          type: string
        message:
          description: Human-readable error message.
          type: string
        details:
          description: Optional extra context (e.g. validation field errors).
          additionalProperties: true

    OTPSendRequest:
      description: Request body for sending an OTP code.
      type: object
      required:
        - email
      properties:
        email:
          description: Email address to send the OTP code to.
          type: string
          format: email

    OTPVerifyRequest:
      description: Request body for verifying an OTP code.
      type: object
      required:
        - email
        - code
      properties:
        email:
          description: Email address the OTP was sent to.
          type: string
          format: email
        code:
          description: The OTP code received via email.
          type: string

    AuthResponse:
      description: Authenticated user information returned after successful login.
      type: object
      required:
        - user_id
        - email
      properties:
        user_id:
          description: Unique identifier for the authenticated user.
          type: integer
          format: int64
        email:
          description: Email address associated with the account.
          type: string

    MessageResponse:
      description: Generic success message.
      type: object
      required:
        - message
      properties:
        message:
          description: Human-readable success message.
          type: string